Cleanup Postgres adapter extension redirection
